Make title selection a method on MangaBakaItem
